Optimize UA SDK structure dissection and centralized data cache
- Optimized recursive mechanism for dissecting OPC UA structures by passing
  const UaExtensionObject & instead of UaVariant, avoiding unnecessary
  encoding/decoding at each level.
- Centralized incoming data cache in ItemUaSdk and removed it from individual
  DataElement objects to reduce memory usage.
- Implemented hierarchical isdirty propagation to optimize assembly of
  structured outgoing data (write path).
- Improved merging logic for LocalizedText and QualifiedName.
- Fixed linker errors in unitTestApp by adding Stats dependency to Makefile.
- Fixed minor compiler warning in open62541 variant.

+++ devOpcuaSup/open62541/DataElementOpen62541Leaf.cpp
@@ -1014,7 +1014,7 @@ DataElementOpen62541Leaf::writeScalar (const char *value, epicsUInt32 len, dbCom
                     }
                 if (ret != 0 && end != value)
                     for (const auto &it: *enumChoices)
-                        if (l == it.first) {
+                        if (l == static_cast<long>(it.first)) {
                             ret = 0;
                             break;
                         }
